Budget
Program compares real expenses with expected expenses. The input file of real expenses: the pay item is a "path in the tree" where node names from the root are separated by a colon and the dates are not in order. The input file of budget (expected expenses): here, expense items are one or more “subtrees” in the general “expenditure tree” (several subtrees are combined with a "+").